# gen/http/key_management/server/server.go
// Code generated by goa v3.21.1, DO NOT EDIT.
//
// key_management HTTP server
//
// Command:
// $ goa gen github.com/Vidalee/FishyKeys/design

package server

import (
	"context"
	"net/http"

	keymanagement "github.com/Vidalee/FishyKeys/gen/key_management"
	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
	"goa.design/plugins/v3/cors"
)

// Server lists the key_management service endpoint HTTP handlers.
type Server struct {
	Mounts          []*MountPoint
	CreateMasterKey http.Handler
	GetKeyStatus    http.Handler
	AddShare        http.Handler
	DeleteShare     http.Handler
	CORS            http.Handler
}

// MountPoint holds information about the mounted endpoints.
type MountPoint struct {
	// Method is the name of the service method served by the mounted HTTP handler.
	Method string
	// Verb is the HTTP method used to match requests to the mounted handler.
	Verb string
	// Pattern is the HTTP request path pattern used to match requests to the
	// mounted handler.
	Pattern string
}

// New instantiates HTTP handlers for all the key_management service endpoints
// using the provided encoder and decoder. The handlers are mounted on the
// given mux using the HTTP verb and path defined in the design. errhandler is
// called whenever a response fails to be encoded. formatter is used to format
// errors returned by the service methods prior to encoding. Both errhandler
// and formatter are optional and can be nil.
func New(
	e *keymanagement.Endpoints,
	mux goahttp.Muxer,
	decoder func(*http.Request) goahttp.Decoder,
	encoder func(context.Context, http.ResponseWriter) goahttp.Encoder,
	errhandler func(context.Context, http.ResponseWriter, error),
	formatter func(ctx context.Context, err error) goahttp.Statuser,
) *Server {
	return &Server{
		Mounts: []*MountPoint{
			{"CreateMasterKey", "POST", "/key_management/create_master_key"},
			{"GetKeyStatus", "GET", "/key_management/status"},
			{"AddShare", "POST", "/key_management/share"},
			{"DeleteShare", "DELETE", "/key_management/share"},
			{"CORS", "OPTIONS", "/key_management/create_master_key"},
			{"CORS", "OPTIONS", "/key_management/status"},
			{"CORS", "OPTIONS", "/key_management/share"},
		},
		CreateMasterKey: NewCreateMasterKeyHandler(e.CreateMasterKey, mux, decoder, encoder, errhandler, formatter),
		GetKeyStatus:    NewGetKeyStatusHandler(e.GetKeyStatus, mux, decoder, encoder, errhandler, formatter),
		AddShare:        NewAddShareHandler(e.AddShare, mux, decoder, encoder, errhandler, formatter),
		DeleteShare:     NewDeleteShareHandler(e.DeleteShare, mux, decoder, encoder, errhandler, formatter),
		CORS:            NewCORSHandler(),
	}
}

// Service returns the name of the service served.
func (s *Server) Service() string { return "key_management" }

// Use wraps the server handlers with the given middleware.
func (s *Server) Use(m func(http.Handler) http.Handler) {
	s.CreateMasterKey = m(s.CreateMasterKey)
	s.GetKeyStatus = m(s.GetKeyStatus)
	s.AddShare = m(s.AddShare)
	s.DeleteShare = m(s.DeleteShare)
	s.CORS = m(s.CORS)
}

// MethodNames returns the methods served.
func (s *Server) MethodNames() []string { return keymanagement.MethodNames[:] }

// Mount configures the mux to serve the key_management endpoints.
func Mount(mux goahttp.Muxer, h *Server) {
	MountCreateMasterKeyHandler(mux, h.CreateMasterKey)
	MountGetKeyStatusHandler(mux, h.GetKeyStatus)
	MountAddShareHandler(mux, h.AddShare)
	MountDeleteShareHandler(mux, h.DeleteShare)
	MountCORSHandler(mux, h.CORS)
}

// Mount configures the mux to serve the key_management endpoints.
func (s *Server) Mount(mux goahttp.Muxer) {
	Mount(mux, s)
}

// MountCreateMasterKeyHandler configures the mux to serve the "key_management"
// service "create_master_key" endpoint.
func MountCreateMasterKeyHandler(mux goahttp.Muxer, h http.Handler) {
	f, ok := HandleKeyManagementOrigin(h).(http.HandlerFunc)
	if !ok {
		f = func(w http.ResponseWriter, r *http.Request) {
			h.ServeHTTP(w, r)
		}
	}
	mux.Handle("POST", "/key_management/create_master_key", f)
}

// NewCreateMasterKeyHandler creates a HTTP handler which loads the HTTP
// request and calls the "key_management" service "create_master_key" endpoint.
func NewCreateMasterKeyHandler(
	endpoint goa.Endpoint,
	mux goahttp.Muxer,
	decoder func(*http.Request) goahttp.Decoder,
	encoder func(context.Context, http.ResponseWriter) goahttp.Encoder,
	errhandler func(context.Context, http.ResponseWriter, error),
	formatter func(ctx context.Context, err error) goahttp.Statuser,
) http.Handler {
	var (
		decodeRequest  = DecodeCreateMasterKeyRequest(mux, decoder)
		encodeResponse = EncodeCreateMasterKeyResponse(encoder)
		encodeError    = EncodeCreateMasterKeyError(encoder, formatter)
	)
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		ctx := context.WithValue(r.Context(), goahttp.AcceptTypeKey, r.Header.Get("Accept"))
		ctx = context.WithValue(ctx, goa.MethodKey, "create_master_key")
		ctx = context.WithValue(ctx, goa.ServiceKey, "key_management")
		payload, err := decodeRequest(r)
		if err != nil {
			if err := encodeError(ctx, w, err); err != nil {
				errhandler(ctx, w, err)
			}
			return
		}
		res, err := endpoint(ctx, payload)
		if err != nil {
			if err := encodeError(ctx, w, err); err != nil {
				errhandler(ctx, w, err)
			}
			return
		}
		if err := encodeResponse(ctx, w, res); err != nil {
			errhandler(ctx, w, err)
		}
	})
}

// MountGetKeyStatusHandler configures the mux to serve the "key_management"
// service "get_key_status" endpoint.
func MountGetKeyStatusHandler(mux goahttp.Muxer, h http.Handler) {
	f, ok := HandleKeyManagementOrigin(h).(http.HandlerFunc)
	if !ok {
		f = func(w http.ResponseWriter, r *http.Request) {
			h.ServeHTTP(w, r)
		}
	}
	mux.Handle("GET", "/key_management/status", f)
}

// NewGetKeyStatusHandler creates a HTTP handler which loads the HTTP request
// and calls the "key_management" service "get_key_status" endpoint.
func NewGetKeyStatusHandler(
	endpoint goa.Endpoint,
	mux goahttp.Muxer,
	decoder func(*http.Request) goahttp.Decoder,
	encoder func(context.Context, http.ResponseWriter) goahttp.Encoder,
	errhandler func(context.Context, http.ResponseWriter, error),
	formatter func(ctx context.Context, err error) goahttp.Statuser,
) http.Handler {
	var (
		encodeResponse = EncodeGetKeyStatusResponse(encoder)
		encodeError    = EncodeGetKeyStatusError(encoder, formatter)
	)
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		ctx := context.WithValue(r.Context(), goahttp.AcceptTypeKey, r.Header.Get("Accept"))
		ctx = context.WithValue(ctx, goa.MethodKey, "get_key_status")
		ctx = context.WithValue(ctx, goa.ServiceKey, "key_management")
		var err error
		res, err := endpoint(ctx, nil)
		if err != nil {
			if err := encodeError(ctx, w, err); err != nil {
				errhandler(ctx, w, err)
			}
			return
		}
		if err := encodeResponse(ctx, w, res); err != nil {
			errhandler(ctx, w, err)
		}
	})
}

// MountAddShareHandler configures the mux to serve the "key_management"
// service "add_share" endpoint.
func MountAddShareHandler(mux goahttp.Muxer, h http.Handler) {
	f, ok := HandleKeyManagementOrigin(h).(http.HandlerFunc)
	if !ok {
		f = func(w http.ResponseWriter, r *http.Request) {
			h.ServeHTTP(w, r)
		}
	}
	mux.Handle("POST", "/key_management/share", f)
}

// NewAddShareHandler creates a HTTP handler which loads the HTTP request and
// calls the "key_management" service "add_share" endpoint.
func NewAddShareHandler(
	endpoint goa.Endpoint,
	mux goahttp.Muxer,
	decoder func(*http.Request) goahttp.Decoder,
	encoder func(context.Context, http.ResponseWriter) goahttp.Encoder,
	errhandler func(context.Context, http.ResponseWriter, error),
	formatter func(ctx context.Context, err error) goahttp.Statuser,
) http.Handler {
	var (
		decodeRequest  = DecodeAddShareRequest(mux, decoder)
		encodeResponse = EncodeAddShareResponse(encoder)
		encodeError    = EncodeAddShareError(encoder, formatter)
	)
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		ctx := context.WithValue(r.Context(), goahttp.AcceptTypeKey, r.Header.Get("Accept"))
		ctx = context.WithValue(ctx, goa.MethodKey, "add_share")
		ctx = context.WithValue(ctx, goa.ServiceKey, "key_management")
		payload, err := decodeRequest(r)
		if err != nil {
			if err := encodeError(ctx, w, err); err != nil {
				errhandler(ctx, w, err)
			}
			return
		}
		res, err := endpoint(ctx, payload)
		if err != nil {
			if err := encodeError(ctx, w, err); err != nil {
				errhandler(ctx, w, err)
			}
			return
		}
		if err := encodeResponse(ctx, w, res); err != nil {
			errhandler(ctx, w, err)
		}
	})
}

// MountDeleteShareHandler configures the mux to serve the "key_management"
// service "delete_share" endpoint.
func MountDeleteShareHandler(mux goahttp.Muxer, h http.Handler) {
	f, ok := HandleKeyManagementOrigin(h).(http.HandlerFunc)
	if !ok {
		f = func(w http.ResponseWriter, r *http.Request) {
			h.ServeHTTP(w, r)
		}
	}
	mux.Handle("DELETE", "/key_management/share", f)
}

// NewDeleteShareHandler creates a HTTP handler which loads the HTTP request
// and calls the "key_management" service "delete_share" endpoint.
func NewDeleteShareHandler(
	endpoint goa.Endpoint,
	mux goahttp.Muxer,
	decoder func(*http.Request) goahttp.Decoder,
	encoder func(context.Context, http.ResponseWriter) goahttp.Encoder,
	errhandler func(context.Context, http.ResponseWriter, error),
	formatter func(ctx context.Context, err error) goahttp.Statuser,
) http.Handler {
	var (
		decodeRequest  = DecodeDeleteShareRequest(mux, decoder)
		encodeResponse = EncodeDeleteShareResponse(encoder)
		encodeError    = EncodeDeleteShareError(encoder, formatter)
	)
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		ctx := context.WithValue(r.Context(), goahttp.AcceptTypeKey, r.Header.Get("Accept"))
		ctx = context.WithValue(ctx, goa.MethodKey, "delete_share")
		ctx = context.WithValue(ctx, goa.ServiceKey, "key_management")
		payload, err := decodeRequest(r)
		if err != nil {
			if err := encodeError(ctx, w, err); err != nil {
				errhandler(ctx, w, err)
			}
			return
		}
		res, err := endpoint(ctx, payload)
		if err != nil {
			if err := encodeError(ctx, w, err); err != nil {
				errhandler(ctx, w, err)
			}
			return
		}
		if err := encodeResponse(ctx, w, res); err != nil {
			errhandler(ctx, w, err)
		}
	})
}

// MountCORSHandler configures the mux to serve the CORS endpoints for the
// service key_management.
func MountCORSHandler(mux goahttp.Muxer, h http.Handler) {
	h = HandleKeyManagementOrigin(h)
	mux.Handle("OPTIONS", "/key_management/create_master_key", h.ServeHTTP)
	mux.Handle("OPTIONS", "/key_management/status", h.ServeHTTP)
	mux.Handle("OPTIONS", "/key_management/share", h.ServeHTTP)
}

// NewCORSHandler creates a HTTP handler which returns a simple 204 response.
func NewCORSHandler() http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		w.WriteHeader(204)
	})
}

// HandleKeyManagementOrigin applies the CORS response headers corresponding to
// the origin for the service key_management.
func HandleKeyManagementOrigin(h http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		origin := r.Header.Get("Origin")
		if origin == "" {
			// Not a CORS request
			h.ServeHTTP(w, r)
			return
		}
		if cors.MatchOrigin(origin, "http://localhost:3000") {
			w.Header().Set("Access-Control-Allow-Origin", origin)
			w.Header().Set("Vary", "Origin")
			w.Header().Set("Access-Control-Max-Age", "3600")
			if acrm := r.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				w.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT, PATCH, DELETE, OPTIONS")
				w.Header().Set("Access-Control-Allow-Headers", "Content-Type")
				w.WriteHeader(204)
				return
			}
			h.ServeHTTP(w, r)
			return
		}
		h.ServeHTTP(w, r)
		return
	})
}